VM/include/PGMInternal.cpp: aidxGuestModeDataTtbr{0,1} need to be wider than uint8_t as the mode table has 514 entries, bugref:3409 bugref:10388
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@110222 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/include/PGMInternal.h
@@ -3785,9 +3785,9 @@ typedef struct PGMCPU
     /** The cached TCR_ELx register (EL0 isn't used). */
     uint64_t                        au64RegTcrEl[3];
     /** Guest mode data table index for a page translation going through TTBR0_ELx (PGM_TYPE_XXX). */
-    uint8_t volatile                aidxGuestModeDataTtbr0[4];
+    uint16_t volatile               aidxGuestModeDataTtbr0[4];
     /** Guest mode data table index for a page translation going through TTBR1_ELx (PGM_TYPE_XXX). */
-    uint8_t volatile                aidxGuestModeDataTtbr1[4];
+    uint16_t volatile               aidxGuestModeDataTtbr1[4];
     /** The initial lookup mask for translations going through TTBR0_ELx. */
     uint64_t                        afLookupMaskTtbr0[4];
     /** The initial lookup mask for translations going through TTBR1_ELx. */
